:recycle: Separate load base from v1/v2 for clarity
